fixed bugs and added delete patient transaction

+++ labsync-patients-microservice/src/main/java/com/labsyncplus/labsync_patients_microservice/Dao/PatientDao.java
@@ -1,9 +1,9 @@
 package com.labsyncplus.labsync_patients_microservice.Dao;
 
-import com.labsyncplus.labsync_patients_microservice.model.Patient;
+import com.labsyncplus.labsync_patients_microservice.model.entity.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PatientDao extends JpaRepository<Patient, Integer> {
+public interface PatientDao extends JpaRepository<Patient, Long> {
 }
